requirements.txt
fastapi==0.101.1
gunicorn==21.2.0
httpx==0.24.1
numpy==1.23.5
pandas==2.0.3
pytest==7.4.0
requests==2.31.0
scikit-learn==1.3.0
uvicorn==0.23.2
hydra-core